Revert "[flutter_tools]: reland android device cleanup (#52542)"
This reverts commit 9186dfc34d2ce982c37e3d3eff89355d088821ab.
