+++ packages/flutter_tools/test/general.shard/android/android_device_test.dart
@@ -166,6 +166,92 @@ void main() {
     });
   });
 
+  group('getAdbDevices', () {
+    MockProcessManager mockProcessManager;
+
+    setUp(() {
+      mockProcessManager = MockProcessManager();
+    });
+
+    testUsingContext('throws on missing adb path', () {
+      final Directory sdkDir = MockAndroidSdk.createSdkDirectory();
+      globals.config.setValue('android-sdk', sdkDir.path);
+
+      final File adbExe = globals.fs.file(getAdbPath(androidSdk));
+      when(mockProcessManager.runSync(
+        <String>[adbExe.path, 'devices', '-l'],
+      )).thenThrow(ArgumentError(adbExe.path));
+      expect(() => getAdbDevices(), throwsToolExit(message: RegExp('Unable to find "adb".*${adbExe.path}')));
+    }, overrides: <Type, Generator>{
+      AndroidSdk: () => MockAndroidSdk(),
+      FileSystem: () => MemoryFileSystem(),
+      ProcessManager: () => mockProcessManager,
+    });
+
+    testUsingContext('throws on failing adb', () {
+      final Directory sdkDir = MockAndroidSdk.createSdkDirectory();
+      globals.config.setValue('android-sdk', sdkDir.path);
+
+      final File adbExe = globals.fs.file(getAdbPath(androidSdk));
+      when(mockProcessManager.runSync(
+        <String>[adbExe.path, 'devices', '-l'],
+      )).thenThrow(ProcessException(adbExe.path, <String>['devices', '-l']));
+      expect(() => getAdbDevices(), throwsToolExit(message: RegExp('Unable to run "adb".*${adbExe.path}')));
+    }, overrides: <Type, Generator>{
+      AndroidSdk: () => MockAndroidSdk(),
+      FileSystem: () => MemoryFileSystem(),
+      ProcessManager: () => mockProcessManager,
+    });
+
+    testUsingContext('physical devices', () {
+      final List<AndroidDevice> devices = <AndroidDevice>[];
+      parseADBDeviceOutput('''
+List of devices attached
+05a02bac               device usb:336592896X product:razor model:Nexus_7 device:flo
+
+''', devices: devices);
+      expect(devices, hasLength(1));
+      expect(devices.first.name, 'Nexus 7');
+      expect(devices.first.category, Category.mobile);
+    });
+
+    testUsingContext('emulators and short listings', () {
+      final List<AndroidDevice> devices = <AndroidDevice>[];
+      parseADBDeviceOutput('''
+List of devices attached
+localhost:36790        device
+0149947A0D01500C       device usb:340787200X
+emulator-5612          host features:shell_2
+
+''', devices: devices);
+      expect(devices, hasLength(3));
+      expect(devices.first.name, 'localhost:36790');
+    });
+
+    testUsingContext('android n', () {
+      final List<AndroidDevice> devices = <AndroidDevice>[];
+      parseADBDeviceOutput('''
+List of devices attached
+ZX1G22JJWR             device usb:3-3 product:shamu model:Nexus_6 device:shamu features:cmd,shell_v2
+''', devices: devices);
+      expect(devices, hasLength(1));
+      expect(devices.first.name, 'Nexus 6');
+    });
+
+    testUsingContext('adb error message', () {
+      final List<AndroidDevice> devices = <AndroidDevice>[];
+      final List<String> diagnostics = <String>[];
+      parseADBDeviceOutput('''
+It appears you do not have 'Android SDK Platform-tools' installed.
+Use the 'android' tool to install them:
+    android update sdk --no-ui --filter 'platform-tools'
+''', devices: devices, diagnostics: diagnostics);
+      expect(devices, hasLength(0));
+      expect(diagnostics, hasLength(1));
+      expect(diagnostics.first, contains('you do not have'));
+    });
+  });
+
   group('parseAdbDeviceProperties', () {
     test('parse adb shell output', () {
       final Map<String, String> properties = parseAdbDeviceProperties(kAdbShellGetprop);
